Shopping Cart functionality
tested the shopping cart adding and removing with the small jar and medium jar.
small jar was basically the test subject for grabbing all the products from the products table and then displaying it on the website (this didn't work).
medium jar was testing the hard-coded method which worked, so I'm going to change every product to make it work.