replacing o2you titles with logo
registration, login and updateprofile page has its o2you title (which also acted as a back function) was replaced by o2you logo image, to keep things as consistent as possible.
Also, with the nav bar, only the text was clickable and not the rest of the button area, and this was due to padding on the list item and not the link item. Shifting the padding to the link item fixes this and the whole area is clickable.