Final Testing Phase I
I am now going through my testing checklist for the final time to evaluate the website and to see if I need to make any changes.
The cart.php page was able to be accessed by users that were not logged in, so using the same code from updateprofile, guests can no longer access the page without first signing in.