Final Test Phase II
This little code on the updateprofile will make it so when the user successfully updates their user information, refreshing the page will refresh the information in the field, and will show the new updated info (as previously it would update, but it wouldnt show the new until it refreshed). This fixes a fail on the testing checklist.