V
The first item that is added to cart will not show the item added to cart message. That's because the cart does not have anything yet and the array needs to be created. Adding the message in that particular line will fix this problem.
Still continuing with the testing document.