VII: Responsive website adjustments
The logo for the website broke when the doctype was re-added. Adjusting the width and height fixed this problem and also fixed the problem where the logo would be bigger than usual on smaller screens.
The welcome message when logged in would also clog the section where the logo sat when the screen was small. Adding the @media with max-wdith (line 30) allows me to turn it off if the width of the screen is smaller than 480px.
Nav bar items weren't completely wrapping, so going one layer deeper and adding flex-wrap solved this, and each nav link will wrap properly instead of being cut off.