VIII: Database
Finished testing.
Added a maxlength for characters for inputs. This is so if a user decides to have an extremely long password etc they will know when they have it the limit of the database which is 255 characters long. Prevents any passwords from getting cutoff without the user knowing.
The updateprofile page also had the title has registration, which is wrong, so it was fixed.