Changes due to DB testing
Max length of each input was changed from 255 to 128 as I feel as the old max number was inappropriate and was too large. This change can be seen on the pages with input.
New SQL file to reflect this change as well.
Flowchart was updated as it forgot a crucial piece of detail: connection script.